refactor: move files into diagramFunctions and split file utils.ts
