+++ src/routes/diagrams/diagramFunctions/blockdiag.ts
