+++ src/routes/diagrams/diagramFunctions/d2.ts
