+++ src/routes/diagrams/diagramFunctions/mermaid.ts
