+++ src/routes/diagrams/diagramFunctions/plantuml.ts
