+++ src/routes/diagrams/diagramFunctions/vega-lite.ts
