+++ src/routes/diagrams/utils.ts
@@ -1,36 +1,3 @@
-import * as pako from 'pako'
-
-export function compressAndEncodeBase64(input: string) {
-  // Convert the input string to a Uint8Array
-  const textEncoder = new TextEncoder()
-  const inputUint8Array = textEncoder.encode(input)
-
-  // Compress the Uint8Array using pako's deflate function
-  const compressedUint8Array = pako.deflate(inputUint8Array) //, { level: 8 });
-
-  // Encode the compressed Uint8Array to a Base64 string
-  const base64Encoded = btoa(
-    String.fromCharCode.apply(null, compressedUint8Array)
-  )
-    .replace(/\+/g, '-')
-    .replace(/\//g, '_')
-
-  return base64Encoded
-}
-
-export function encodeBase64(input: string) {
-  // Convert the input string to a Uint8Array
-  const textEncoder = new TextEncoder()
-  const inputUint8Array = textEncoder.encode(input)
-
-  // Encode the compressed Uint8Array to a Base64 string
-  const base64Encoded = btoa(String.fromCharCode.apply(null, inputUint8Array))
-    .replace(/\+/g, '-')
-    .replace(/\//g, '_')
-
-  return base64Encoded
-}
-
 export const diagramLanguages = [
   "blockdiag",
   "bpmn",
